requirements.txt
# Web Frameworks and UI
Flask
Flask-Cors
streamlit

# PDF and Image Processing
PyPDF2
Pillow
pdf2image
pytesseract

# Data Scraping and HTTP Requests
beautifulsoup4
requests
tqdm

# Data Handling
pandas

# API Clients and Helpers
supabase
tenacity

# AI-Agent
flask
flask-cors
requests
python-dotenv

# MCP
google.generativeai
sentence_transformers
supabase
mcp
